Say what a node index is instead of what it converts to
The index type was asked for in From<usize> and Into<usize>, which
between them admit usize and nothing else, so every tree carried a
pointer-sized index whatever it was indexing. A node is 184 bytes and 8
of them are the cell length; the other 168 are six facets and eight
children, all of them indices. Narrowing one to a u32 halves the node,
where narrowing the cell length saved four percent.

Slot says what an index is: a slot it can count back to, and one it
makes from a node count or refuses. Trees are built with u32 indices
now, which reach four billion nodes, some four hundred gigabytes of
tree.

Subdividing takes a usize. It used to take an index, so a caller
holding a count had to make one, and making one is the fallible part:
minting the children is the only moment a tree can outgrow what its
indices reach. That is where it is checked, and the only place it can
be.

Measured on a bone at scale nine: 43785 nodes, 8.06 MB of them before
and 4.20 MB after, at the same speed and the same mesh.

+++ src/geometry/grid/defeature/mod.rs
@@ -9,7 +9,7 @@ use std::hash::Hash;
 
 impl<V: Copy + Eq + Hash> Pixels<V> {
     pub fn defeature(self, minimum: usize) -> Result<Self, &'static str> {
-        let mut quadtree = Quadtree::<u16, usize, V>::try_from(self)?;
+        let mut quadtree = Quadtree::<u16, u32, V>::try_from(self)?;
         quadtree.defeature(minimum);
         Ok(Self::from(&quadtree))
     }
@@ -17,7 +17,7 @@ impl<V: Copy + Eq + Hash> Pixels<V> {
 
 impl<V: Copy + Eq + Hash> Voxels<V> {
     pub fn defeature(self, minimum: usize) -> Result<Self, &'static str> {
-        let mut octree = Octree::<u16, usize, V>::try_from(self)?;
+        let mut octree = Octree::<u16, u32, V>::try_from(self)?;
         octree.defeature(minimum);
         Ok(Self::from(&octree))
     }
